refactor(frontend): decompose AskView and ArticleReader monolith components (#710)
Extract hooks and utilities from two oversized components:

AskView (310→90 lines): extract useAskStream, useFileBack, useExportConversation
ArticleReader (407→170 lines): extract markdownComponents, preprocessMarkdown, useArticleEditor
